requirements.txt
pvporcupine>=3.0.0
pvrecorder>=1.2.0
faster-whisper>=1.0.0
anthropic>=0.40.0
edge-tts>=6.1.0
pygame>=2.6.0
pystray>=0.19.0
Pillow>=10.0.0
python-dotenv>=1.0.0
